ml_service/requirements.txt
feedparser
sentence-transformers
chromadb
transformers
yfinance
ta
pandas
numpy
requests
torch
python-dotenv
huggingface_hub
